+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.1.0-beta.19 - [November 9, 2024](https://github.com/lando/vitepress-theme-default-plus/releases/tag/v1.1.0-beta.19)
 
+* Added `frontmatter.url-loader` to allow `.md` files to pull content from remote `markdown` sources
 * Added `version` alias information to config
 * Fixed bug preventing user specified `buildEnd` and `transformPageData` from running after theme's
 * Fixed bug preventing `mvb` from correctly setting the `mvbase`